Avoid fn_proxy cache for load_deferred_integration
We hit a segmentation fault here once and reliably enabled/disabled
it and we'd watch the segfault come and go away. When trying to
reproduce it we couldn't recreate the issue. There is now a
customer report of a segmentation fault on Redis::__construct, which
is a deferred loading trigger. Given the prior investigation I am
disabling this cache as a safety precaution.