[ddprof] Support v3 intake format; HTTP bugfixes (#1175)
* Support v3 intake format; HTTP bugfixes

This should support the v3 intake format.
Also force the HTTP layer to
 * use Connection: close and do tail-closing for HTTP reqs
 * properly handle SIGPIPE (by not allowing it to be thrown!)

* Fix manual merge artifacts

* Various HTTP fixes

 - dd_send no longer defaults runtime, since this is not useful for v3
 - fix typo; s/owernship/ownership/g
 - change HTTP route if apikey is specified
 - http.c should support both macos and linux now

* Better safeguarding in dd_send

During a send-type operation, parameters are checked to see whether they
are required.  Before, if a parameter was required, an error was thrown
only when that parameter also had no default value (why would you have a
default value for a required paramater, you may ask--nearsightedness is
why)

This patch very simply overrides the default column to specify an error
to be given when the value is absent.  It also slightly restructures the
order of defaulting in the handler that populates the HTTP header.

* Revert "Better safeguarding in dd_send"

This reverts commit ea03f6c158b6adbe74d350796002b2bedab777cf.

Co-authored-by: sanchda <sanchda@users.noreply.github.com>
Co-authored-by: Levi Morrison <levi.morrison@datadoghq.com>